Revamped the whole app, leverage more of server components advantages and fixed some bugs.

+++ src/actions/pixelActions.ts
@@ -1,12 +1,13 @@
 "use server";
 import db from "@/lib/db";
 import { Pixel } from "@prisma/client";
-import { getSessionUserId, handleServerError, normalizeDate } from "./actionUtils";
+import { getSessionUserId, logServerError, normalizeDate } from "./actionUtils";
+import { cacheTag } from "next/dist/server/use-cache/cache-tag";
+import { revalidateTag } from "next/cache";
 
-export const getUserPixels = async (): Promise<Pixel[]> => {
+export const getUserPixels = async (userId: string): Promise<ServerActionResponse<Pixel[]>> => {
+    'use cache'
     try {
-        const userId = await getSessionUserId();
-
         const pixels = await db.pixel.findMany({
             where: {
                 userId,
@@ -28,23 +29,30 @@ export const getUserPixels = async (): Promise<Pixel[]> => {
                 },
             },
         });
+        cacheTag(`p`);
 
-        return pixels.map(pixel => ({
-            ...pixel,
-            mood: pixel.moods.length > 0 ? pixel.moods[0].mood : null, // Adapt mood data
-        }));
+        cacheTag(`pixels-${userId}`);
+        return { success: true, data: pixels }
+        // return pixels.map(pixel => ({
+        //     ...pixel,
+        //     mood: pixel.moods.length > 0 ? pixel.moods[0].mood : null, // Adapt mood data
+        // }));
     } catch (error: unknown) {
-        handleServerError(error as Error, "retrieving pixels.");
-        return [];
+        logServerError(error as Error, 'editing user event in db');
+        return {
+            success: false,
+            error: 'Failed to fetch user pixels. Please try again later.',
+        };
     }
 };
 
 export const getUserPixelsByRange = async (
     from: Date,
-    to?: Date
-): Promise<Pixel[]> => {
+    to?: Date,
+    userId?: string
+): Promise<ServerActionResponse<Pixel[]>> => {
+    'use cache'
     try {
-        const userId = await getSessionUserId();
 
         const fromDate = new Date(
             Date.UTC(from.getFullYear(), from.getMonth(), from.getDate())
@@ -80,34 +88,46 @@ export const getUserPixelsByRange = async (
                 },
             },
         });
-        return pixels;
+        cacheTag(`p`);
+
+        return { success: true, data: pixels };
 
     } catch (error: unknown) {
-
-        handleServerError(
-            error as Error,
-            `retrieving pixels by range .`
-        );
-        return [];
+        logServerError(error as Error, 'editing pixel event in db');
+        return {
+            success: false,
+            error: `Failed to fetch user pixels by range ${from} to ${to}. Please try again later.`,
+        };
     }
 };
 
-export const upsertUserPixel = async (pixel: Pixel, eventIds: string[] = [], moodIds: string[] = []): Promise<Pixel | null> => {
-    try {
-        const userId = await getSessionUserId();
-        const normalizedDate = normalizeDate(pixel.pixelDate);
+export const upsertUserPixel = async (pixel: Pixel, eventIds: string[] = [], moodIds: string[] = []): Promise<ServerActionResponse<Pixel>> => {
 
-        // Use Prisma's upsert function
+    const fetchUserId = await getSessionUserId();
+    if (!fetchUserId.success) {
+        return { success: false, error: 'User not authenticated.' };
+    }
+    const userId = fetchUserId.data;
+
+    let normalizedDate: Date;
+    try {
+        normalizedDate = normalizeDate(pixel.pixelDate);
+    } catch (error) {
+        logServerError(error as Error, 'normalizing date for pixel');
+        return { success: false, error: 'Failed to normalize date.' };
+    }
+    try {
+
         const upsertedPixel = await db.pixel.upsert({
             where: {
-                userId_pixelDate: { // Compound unique index name.  Must match what's in the Prisma schema
+                userId_pixelDate: {
                     userId: userId,
                     pixelDate: normalizedDate,
                 },
             },
             update: {
                 events: {
-                    deleteMany: {}, // Disconnect all existing events
+                    deleteMany: {},
                     create: eventIds.map((eventId) => ({
                         event: {
                             connect: {
@@ -117,8 +137,8 @@ export const upsertUserPixel = async (pixel: Pixel, eventIds: string[] = [], moo
                     })),
                 },
                 moods: {
-                    deleteMany: {}, // Disconnect all existing moods
-                    create: moodIds.map((moodId) => ({ // Connect mood
+                    deleteMany: {},
+                    create: moodIds.map((moodId) => ({
                         mood: {
                             connect: {
                                 id: moodId,
@@ -127,7 +147,7 @@ export const upsertUserPixel = async (pixel: Pixel, eventIds: string[] = [], moo
                     })),
                 },
             },
-            create: { // Create a new record.
+            create: {
                 pixelDate: normalizedDate,
                 userId: userId,
                 events: {
@@ -179,13 +199,58 @@ export const upsertUserPixel = async (pixel: Pixel, eventIds: string[] = [], moo
             }
         })
 
-        return {
-            ...returnPixel,
-            mood: returnPixel && returnPixel.moods.length > 0 ? returnPixel.moods[0].mood : null, // Adapt mood data
-        };
+        revalidateTag(`pixels-${userId}`);
+        revalidateTag(`pixel-${userId}-${normalizedDate}`);
+        revalidateTag(`pixels-${userId}-${pixel.pixelDate}`);
+        revalidateTag(`p`);
+        return { success: true, data: returnPixel }
+        // return {
+        //     ...returnPixel,
+        //     mood: returnPixel && returnPixel.moods.length > 0 ? returnPixel.moods[0].mood : null, // Adapt mood data
+        // };
 
     } catch (error: unknown) {
-        handleServerError(error as Error, "adding pixel for user.");
-        return null;
+        logServerError(error as Error, 'adding pixel in db');
+        return {
+            success: false,
+            error: 'Failed to add pixel. Please try again later.',
+        };
+    }
+};
+
+export const getUserPixelsByYear = async (year: number, userId: string): Promise<ServerActionResponse<Pixel[]>> => {
+    'use cache'
+    try {
+        const pixels = await db.pixel.findMany({
+            where: {
+                userId,
+                pixelDate: {
+                    gte: new Date(year, 0, 1),
+                    lte: new Date(year, 11, 31, 23, 59, 59, 999),
+                },
+            },
+            include: {
+                moods: { //Include mood via the join table
+                    include: {
+                        mood: true,
+                    },
+                },
+                events: {
+                    include: {
+                        event: true,
+                    },
+                },
+            },
+        });
+        cacheTag(`p`);
+
+        cacheTag(`pixels-${userId}-${year}`);
+        return { success: true, data: pixels };
+    } catch (error: unknown) {
+        logServerError(error as Error, 'fetching user pixels by year');
+        return {
+            success: false,
+            error: `Failed to fetch pixels by year ${year}. Please try again later.`,
+        };
     }
 };
